added document routes and logic

+++ backend/controller/document_controller.js
@@ -1,40 +1,79 @@
+const document = require("../Model/Document");
 const Document = require("../Model/Document");
 const Tag  = require("../Model/Tag");
-
+const User = require("../Model/user.js")
 module.exports.createdoc = async (req, res) => {
-  const title = req.body.name;
-  const author = req.body.author;
+  const title = req.body.title;
+  console.log(title);
+
+  // this is id of user who is creating document
+
+  const id = req.body.id;
   const description = req.body.description;
+  // check if user is valid
+  console.log(req.body)
+
+    const user = await User.findById(id);
+    // we will get object directly here 
+
+ 
+    if(!user){
+      // invalid user
+      res.status(400).json({
+        message:'not a valid user as user is not found in database'
+      })
+
+      return;
+    }
   const newDoc = await Document.create({
-    title,
-    author,
+    title: title,
+    createdBy: id,
     description,
   });
   res.status(200).json({
     data: newDoc,
   });
+ 
 };
 
 
 module.exports.showAllDocs = async (req, res) => {
+ try{
   const id = req.params.id;
   
-  const docs = await Document.findById(id);
+  const docs = await Document.find({createdBy:id}).populate('createdBy');
 
   res.status(200).json({
     data: docs,
   });
+ }catch(e){
+    console.log(e);
+ }
+ 
 };
 
 
-modules.exports.deleteDoc = async (req, res) => {
-  const id = req.params.id;
-  c
-
-  const doc = await Document.findByIdAndDelete(id);
+module.exports.deleteDoc = async (req, res) => {
+  
+  try{
+    const id = req.params.id;
+      const doc = await Document.findById(id);
+   console.log(doc);
+      if(!doc){
+        res.status(400).json({
+          message:'doc not found'
+        });
+        return;
+      }
+ 
+     const deleted_doc = await Document.findByIdAndDelete(id);
   res.status(200).json({
     data: doc,
   });
+  }catch(e){
+    console.log(e);
+  }
+ 
 };
 
 
@@ -43,34 +82,53 @@ module.exports.updateDoc = async (req, res) => {
   // user id
     
   try {
-  
-    const id = req.params.id;
+    const title = req.body.title;
+    console.log(title);
+    // this is id of the document
+    const id = req.body.id;
+    const description = req.body.description;
 
-    const doc = await Document.findById(id);
-
-    doc.approve = !doc.approve;
-
-    doc.save();
+    const doc = await document.findById(id);
+    console.log(doc);
+    if (!doc) {
+        res.status(400).json({
+          message:'doc not found'
+        })
+    }
+    
+    doc.title = title;
+    doc.description = description;
+        doc.save();
+     res.status(200).json({
+      data: doc,
+     })
   } catch (e) {
-    req.status(404).json({
-      message: "Doc is not found",
-    });
+      console.log(e);
   }
 
-  res.status(200).json({
-    data: doc,
-  });
+  
 };
 
 
 module.exports.approveDoc = async (req, res) => {
-  const id = req.params.id;
-  const doc = await Document.findByIdAndUpdate(id, {
-    approved: true,
-  });
-  res.status(200).json({
-    data: doc,
-  });
+  try{
+    const id = req.params.id;
+    const doc = await Document.findById(id);
+    if(!doc){
+      res.status(400).json({
+        message:'doc not found'
+      })
+      return;
+    }
+    doc.approved = !doc.approved;
+    doc.save();
+    res.status(200).json({
+      data: doc,
+    });
+  }catch(e){
+    console.log(e);
+  }
+ 
 };
